CR-4350 Forward port DOCS-194: Add option to allow merge of generated and docbkx source directories

+++ forgerock-doc-maven-plugin/trunk/src/main/java/org/forgerock/doc/maven/pre/ModifiableCopy.java
@@ -52,7 +52,9 @@ public class ModifiableCopy {
         final File outputDir = m.getDocbkxModifiableSourcesDirectory();
 
         try {
-            FileUtils.copyDirectory(sourceDir, outputDir);
+            if (m.doOverwriteModifiableCopy()) {
+                FileUtils.copyDirectory(sourceDir, outputDir);
+            }
         } catch (IOException e) {
             throw new MojoExecutionException("Failed to copy sources", e);
         }
